requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
asyncmy
asyncpg
bcrypt
fastapi==0.109.2
greenlet==1.1.2
gunicorn==20.1.0
httpx
passlib==1.7.4
pycryptodome
pydantic>=2.6.1
pydantic[email]
pytest
pytest-asyncio
python-jose[cryptography]
python-multipart
SQLAlchemy==2.0.27
sqlmodel>=0.0.14
uvicorn>=0.20.0
